requirements.txt
DomiKnowS==0.533
peft==0.10.0
pandas==2.2.1
numpy==1.26.4
sentencepiece==0.2.0
torch==2.2.1
tqdm==4.66.2
transformers==4.39.2
spacy==3.7.4
accelerate==0.29.1
python-dotenv